AI based on Fitness Proportionate Probability
Added a more intelligent AI implementation which selects moves according
to the amount of damage.
It uses Fitness Proportionate Selection in which the moves are weighted
according to damage, such that the higher damaging moves have a higher
probability of being chosen.
This will seem more intelligent than the random AI, however will not
pick a perfect move every time, meaning it can be beaten with a perfect
strategy.