Training via GUI
Training an AI can now be controlled via the GUI.
Currently there exist controls for creating a new training session,
starting the session and pausing the session.
Due to the async nature of the starting and stopping of the evolutionary
algorithm, internal state controls when certain actions are allowed to
be actioned by the user.