Trainer.Console project
Moved the console trainer to a separate assembly. The main purpose of
this is to remove the .exe being included in the GUI assembly, as it
isn't needed. The Trainer project is now a class library assembly.
Improvements have been made to the console app, including the ability to
pause / resume training.