Average fitness calculation
Altered the way the average fitness of a genome is calculated.
Previously, the average was the fitness score divided by the total
number of games played, which was equivalent to the number of game
combinations multiplied by the number of iterations for each game
combination.
Now, each game combination will calculate its own local average, which
consists of the total fitness divided by the number of iterations that
occurred for that particular combination.